move threads, forksafe, and _unpatched to internal/utils
